Nhom3 an

+++ Source/app/src/main/java/com/laptrinhdidong/nhom3/quanlichitieu/Object/Nhom3QuocItemBarChart.kt
@@ -1,4 +1,6 @@
 package com.laptrinhdidong.nhom3.quanlichitieu.Object
 
-class Nhom3QuocItemBarChart(var title_bc: String, var money_collect: Long, var money_lost: Long, var total_money: Long) {
+import java.math.BigDecimal
+
+class Nhom3QuocItemBarChart(var title_bc: String, var money_collect: BigDecimal, var money_lost: BigDecimal, var total_money: BigDecimal) {
 }
